+++ tests/test_axis.py
@@ -7,7 +7,6 @@ from numpy.testing import assert_allclose, assert_array_equal
 from pytest import approx
 
 import boost_histogram as bh
-import boost_histogram.utils
 
 
 @pytest.mark.parametrize(